requirements.txt
anyio==4.7.0
appdirs==1.4.4
argcomplete==3.5.2
certifi==2024.12.14
charset-normalizer==3.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_resources==6.4.5
pillow==11.0.0
pyTelegramBotAPI==4.25.0
python-barcode==0.15.1
python-escpos==3.1
python-telegram-bot==21.9
PyYAML==6.0.2
qrcode==8.0
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
telebot==0.0.5
urllib3==2.3.0
